+++ tests/components/AddItemForm/AddItemForm.test.jsx
@@ -1,0 +1,11 @@
+// import { render } from '@testing-library/react';
+// import { AddItemForm } from '../../../src/components/AddItemForm/AddItemForm';
+
+describe('AddItemForm', () => {
+	it('renders', () => {
+		// needs to mock data
+		// needs to mock listPath
+		// needs to mock firebase call
+		// render(<AddItemForm />);
+	});
+});
